feat: 4.15. рефакторинг генерации публикаций
