+++ src/utils/getRandomInt.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const getRandomInt = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min)) + min;
+};
+
+module.exports = getRandomInt;
