+++ src/utils/index.js
@@ -2,8 +2,20 @@
 
 const readFileContent = require(`./readFileContent`);
 const generateArray = require(`./generateArray`);
+const getRandomInt = require(`./getRandomInt`);
+const getCreatedDate = require(`./getCreatedDate`);
+const shuffle = require(`./shuffle`);
+const getPostTitle = require(`./getPostTitle`);
+const getPostCategories = require(`./getPostCategories`);
+const getText = require(`./getText`);
 
 module.exports = {
   readFileContent,
   generateArray,
+  getRandomInt,
+  getCreatedDate,
+  shuffle,
+  getPostTitle,
+  getPostCategories,
+  getText,
 };
