+++ src/utils/shuffle.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const shuffle = (someArray) => {
+  for (let i = someArray.length - 1; i > 0; i--) {
+    const randomPosition = Math.floor(Math.random() * i);
+    [someArray[i], someArray[randomPosition]] = [someArray[randomPosition], someArray[i]];
+  }
+
+  return someArray;
+};
+
+module.exports = shuffle;
